:bug: Fix: comment API 충돌 해결
Fix: comment API 충돌 해결
