+++ app/_apis/schema/projectResponse.ts
@@ -1,31 +1,3 @@
-export interface ProjectResponseType {
-  content: ProjectData[];
-  customPageable: CustomPageable;
-}
-
-export interface ProjectData {
-  projectId: number;
-  thumbnailUrl: string;
-  stackList: string[];
-  likeCount: number;
-  isLiked: boolean;
-  projectTitle: string;
-  introduction: string;
-  viewCount: number;
-  createdAt: string;
-  modifiedAt: string;
-}
-
-export interface CustomPageable {
-  first: boolean;
-  last: boolean;
-  hasNext: boolean;
-  totalPages: number;
-  totalElements: number;
-  page: number;
-  size: number;
-}
-
 export interface ProjectResponse {
   projectId: number;
   memberId: number;
@@ -38,43 +10,56 @@ export interface ProjectResponse {
   introductions: string;
   serviceUrl: string;
   imageType: string;
-  projectLinks: [
-    {
-      id: number;
-      siteType: string;
-      url: string;
-    },
-  ];
-  projectTechStacks: [
-    {
-      id: number;
-      techStack: string;
-    },
-  ];
+  projectLinks: ProjectLinkList[];
+  projectTechStacks: TechStackList[];
   thumbnailUrl: string;
-  imageUrlList: [
-    {
-      id: number;
-      url: string;
-      idx: number;
-    },
-  ];
+  imageUrlList: ImageUrlList[];
   isMine: boolean;
   isLiked: false;
 }
 
+export interface EditProjectResponse {
+  title: string;
+  introduction: string;
+  content: string;
+  serviceUrl: string;
+  imageType: string;
+  projectLinks: ProjectLinkList[];
+  projectTeammates: ProjectTeammateList[];
+  techStacks: TechStackList[];
+  thumbnailUrl: string;
+  imageUrlList: string[];
+}
+
+interface ProjectLinkList {
+  id: number;
+  siteType: string;
+  url: string;
+}
+
+interface ProjectTeammateList {
+  id: number;
+  teammateName: string;
+  job: string;
+  url: string;
+}
+
+interface TechStackList {
+  id: number;
+  techStack: string;
+}
+
+interface ImageUrlList {
+  id: number;
+  url: string;
+  idx: number;
+}
+
 export interface TeamMemberResponse extends Array<TeamMember> {}
 
-export interface TeamMember {
+interface TeamMember {
   job: string;
-  teammateList: [
-    {
-      id: number;
-      teammateName: string;
-      job: string;
-      url: string;
-    },
-  ];
+  teammateList: ProjectTeammateList[];
 }
 
 export interface TotalRatingResponse {
