+++ app/_queryFactory/projectListQuery.ts
@@ -1,12 +1,13 @@
 import { createQueryKeys } from "@lukemorales/query-key-factory";
-import { GetMyProjectListParams, projectListAPI } from "../_apis/projectListAPI";
+import { projectListApi } from "../_apis/projectListApi";
 import { MyPageProjectListType } from "../profile/[userId]/_components/MypageProjectSection";
+import { GetMyProjectListParams } from "../_types/ProjectListType";
 
 export const profileProjectListKeys = createQueryKeys("projectList", {
   profileList: ({ page, size, userId }: GetMyProjectListParams, selectDataType: MyPageProjectListType) => ({
     queryKey: [`${selectDataType}List`, page],
     queryFn: async () =>
-      await projectListAPI.getMyProjectList({ page: page, size: size, userId: userId }, selectDataType),
+      await projectListApi.getMyProjectList({ page: page, size: size, userId: userId }, selectDataType),
   }),
 });
 
@@ -14,6 +15,6 @@ export const NewProjectListKeys = createQueryKeys("project", {
   profileList: ({ page, size, userId }: GetMyProjectListParams, selectDataType: MyPageProjectListType) => ({
     queryKey: [`${selectDataType}List`, page],
     queryFn: async () =>
-      await projectListAPI.getMyProjectList({ page: page, size: size, userId: userId }, selectDataType),
+      await projectListApi.getMyProjectList({ page: page, size: size, userId: userId }, selectDataType),
   }),
 });
