--- app/_types/ProjectListDataType.ts
@@ -1,8 +1,0 @@
-export interface ProjectListParams {
-  page?: number;
-  size?: number;
-  limit?: number;
-  searchString?: string;
-  projectTechStacks?: string[];
-  sortCondition?: "LIKES" | "VIEWS" | "RECENT";
-}
